Move Kafka transport into sei-db/queue/kafka
Pull the Kafka plumbing out of the state-store offload package into a
standalone queue transport library:

- writer construction (WriterConfig, SASL validation and mechanisms,
  the AWS MSK IAM signer) moves from offload,
- reader construction (ReaderConfig) moves from offload/consumer,
- the generic batching consume loop (Consumer, Options, Sink) and its
  throughput/lag metrics move from offload/consumer.

The loop is now payload-agnostic: Sink.WriteBatch receives raw Kafka
messages and owns decoding. The Bigtable sink decodes changelog entries
itself, so offload/consumer keeps only the changelog-to-Bigtable
application. offload.KafkaConfig stays as a type alias so existing
callers keep compiling. Meter and metric names are unchanged.
