+++ sei-db/queue/kafka/metrics_test.go
@@ -1,18 +1,18 @@
-package consumer
+package kafka
 
 import (
 	"context"
 	"testing"
 	"time"
 
-	"github.com/segmentio/kafka-go"
+	kafkago "github.com/segmentio/kafka-go"
 	"github.com/stretchr/testify/require"
 )
 
 func TestBatchLag(t *testing.T) {
 	// Max over the batch of (HighWaterMark - Offset - 1); the head message
 	// (offset == HWM-1) has lag 0.
-	msgs := []kafka.Message{
+	msgs := []kafkago.Message{
 		{Offset: 10, HighWaterMark: 100}, // 89
 		{Offset: 98, HighWaterMark: 100}, // 1
 		{Offset: 99, HighWaterMark: 100}, // 0 (head)
@@ -20,7 +20,7 @@ func TestBatchLag(t *testing.T) {
 	require.Equal(t, int64(89), batchLag(msgs))
 
 	// No watermark information -> sentinel -1 so the gauge is left untouched.
-	require.Equal(t, int64(-1), batchLag([]kafka.Message{{Offset: 5}}))
+	require.Equal(t, int64(-1), batchLag([]kafkago.Message{{Offset: 5}}))
 	require.Equal(t, int64(-1), batchLag(nil))
 }
